bdd/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
pyyaml==6.0.1
jinja2==3.1.3
wiremock==3.0.2
pytest==8.0.2